requirements.txt
fastapi=0.92.0
uvicorn=0.20.0
python-dotenv=1.0.0
databases=0.7.0
asyncpg=0.27.0
SQLAlchemy=1.4.46

